Break out more from S3ReservationsStorage
